requirements.txt
flet==0.22.0
idna==3.7
certifi==2024.2.2
urllib3==2.2.1
charset-normalizer==3.3.2
chardet==5.2.0
PySocks==1.7.1
opencv-python
pillow
numpy
socks
security